+++ agate/table/print_html.py
@@ -5,14 +5,16 @@ import sys
 import six
 
 
-def print_html(self, max_rows=None, max_columns=None, output=sys.stdout):
+def print_html(self, max_rows=20, max_columns=6, output=sys.stdout):
     """
     Print an HTML version of this table.
 
     :param max_rows:
-        The maximum number of rows to display before truncating the data.
+        The maximum number of rows to display before truncating the data. By
+        default all rows will be printed.
     :param max_columns:
-        The maximum number of columns to display before truncating the data.
+        The maximum number of columns to display before truncating the data. By
+        default all columns will be printed.
     :param output:
         A file-like object to print to. Defaults to :code:`sys.stdout`.
     """