+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 1.3.1
 -----
 
+* :meth:`.Table.print_table` now defaults to printing ``20`` rows and ``6`` columns. (#589)
 * Added Eli Murray to AUTHORS.
 * :meth:`.Table.__init__` now accepts a dict to specify partial column types. (#580)
 * :meth:`.Table.from_csv` now accepts a ``skip_lines`` argument. (#581)